backend/requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar

# index.py: 1
aiohttp == 2.3.6

# index.py: 2
python_socketio == 1.8.4
keras
tensorflow
pandas
h5py
